Reorg file structure
